blender/pyproject.toml
[project]
name = "zhottem-blender"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fake-bpy-module>=20250326",
]
